refactor(module): move up static Instance methods and move down unrelated
