+++ lua/banana/utils/debug.lua
@@ -17,4 +17,22 @@ function M.isdev()
     return ok and val
 end
 
+function M.spam()
+    local testFile = [[
+			.asdf {
+				hl-bg: #0000ff;
+			}
+		]]
+    local startTime = vim.uv.hrtime()
+    for _ = 1, 1000 do
+        local _ = require('banana.ncss.parser').parseTextSlow(testFile)
+    end
+    vim.notify((vim.uv.hrtime() - startTime) / 1e6 .. "ms\n")
+    startTime = vim.uv.hrtime()
+    for _ = 1, 1000 do
+        local _ = require('banana.ncss.parser').parseText(testFile)
+    end
+    vim.notify((vim.uv.hrtime() - startTime) / 1e6 .. "ms\n")
+end
+
 return M
